+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## ## 1.0.3-beta + [UPDATES] - 2025 XXX
+- Removed per-command --strict flags.
+- Added --strict as global option, to enable parsing in strict mode.
+- Added -q, --quiet as global option, to reduce output (show only errors).
+- Added --silent as global option, will suppress all output (even errors, exit code only).
+
 ## 1.0.3-beta - 2025 Sep
 - **Updated:** Now uses latest `yini-parser` library `v1.1.0-beta`, for greatly improved parsing, compatibility, and error handling with more descriptive and accurate error messages.